feat: wps 和 ms office 的逻辑拆分，使逻辑更清晰易懂
