+++ crates/office-converter/src/lib.rs
@@ -1,6 +1,9 @@
+mod com_utils;
 pub mod converter;
 pub mod detector;
 pub mod error;
+mod ms_office;
+mod wps_office;
 
 pub use converter::{convert_to_html, convert_to_html_with_options, ConvertOptions};
 pub use detector::{detect_office_apps, get_preferred_office, OfficeApp, OfficeInfo};
